+++ crates/jwk/Cargo.toml
@@ -11,3 +11,4 @@ base64 = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 sha2 = "0.10.8"
+thiserror = { workspace = true }
