+++ crates/keys/src/key_manager/key_store/mod.rs
@@ -1,7 +1,9 @@
 pub mod in_memory_key_store;
 
-use crate::key::{Curve, KeyError, PublicKey};
+use crate::key::{KeyError, PublicKey};
+use crypto::{CryptoError, Curve, Signer};
 use josekit::jws::JwsSigner;
+use jwk::JwkError;
 use std::sync::Arc;
 
 #[derive(thiserror::Error, Debug, Clone, PartialEq)]
@@ -12,6 +14,10 @@ pub enum KeyStoreError {
     KeyError(#[from] KeyError),
     #[error("key not found {0}")]
     KeyNotFound(String),
+    #[error(transparent)]
+    CryptoError(#[from] CryptoError),
+    #[error(transparent)]
+    JwkError(#[from] JwkError),
 }
 
 // Trait for storing and retrieving private keys.
@@ -22,5 +28,5 @@ pub trait KeyStore: Send + Sync {
     fn get_all_aliases(&self) -> Result<Vec<String>, KeyStoreError>;
     fn sign(&self, key_alias: &str, payload: &[u8]) -> Result<Vec<u8>, KeyStoreError>;
     fn get_public_key(&self, key_alias: &str) -> Result<Arc<dyn PublicKey>, KeyStoreError>;
-    fn get_jws_signer(&self, key_alias: &str) -> Result<Arc<dyn JwsSigner>, KeyStoreError>;
+    fn get_signer(&self, key_alias: &str) -> Result<Signer, KeyStoreError>;
 }
